refactor: fix more code smells
